refactor: change directory name
- model 패키지를 domain으로 통일
