+++ src/main/java/com/kakaoseventeen/dogwalking/chat/repository/ChatMessageRepository.java
@@ -1,6 +1,6 @@
 package com.kakaoseventeen.dogwalking.chat.repository;
 
-import com.kakaoseventeen.dogwalking.chat.model.ChatMessage;
+import com.kakaoseventeen.dogwalking.chat.domain.ChatMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
