+++ src/main/java/com/kakaoseventeen/dogwalking/chat/repository/ChatRoomRepository.java
@@ -1,6 +1,6 @@
 package com.kakaoseventeen.dogwalking.chat.repository;
 
-import com.kakaoseventeen.dogwalking.chat.model.ChatRoom;
+import com.kakaoseventeen.dogwalking.chat.domain.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
